@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@editorjs/editorjs": "^2.30.2",
+    "@editorjs/editorjs": "^2.30.4",
     "@editorjs/header": "^2.8.7",
     "@editorjs/list": "^1.10.0",
     "@hookform/resolvers": "^3.9.0",
@@ -41,39 +41,45 @@
     "@radix-ui/react-toggle": "^1.1.0",
     "@radix-ui/react-toggle-group": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.2",
+    "@tanstack/react-query": "^5.51.23",
+    "@tanstack/react-query-devtools": "^5.51.23",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
     "embla-carousel-react": "^8.1.8",
     "input-otp": "^1.2.4",
+    "jotai": "^2.9.2",
+    "lodash.template": "^4.5.0",
     "lucide-react": "^0.424.0",
     "next": "14.2.5",
     "next-themes": "^0.3.0",
-    "react": "^18",
+    "nextjs-toploader": "^1.6.12",
+    "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
-    "react-dom": "^18",
+    "react-dom": "^18.3.1",
     "react-editor-js": "^2.1.0",
     "react-hook-form": "^7.52.2",
-    "react-resizable-panels": "^2.0.22",
+    "react-resizable-panels": "^2.0.23",
     "react-textarea-autosize": "^8.5.3",
     "recharts": "^2.12.7",
     "sonner": "^1.5.0",
     "swr": "^2.2.5",
-    "tailwind-merge": "^2.4.0",
+    "tailwind-merge": "^2.5.0",
     "tailwindcss-animate": "^1.0.7",
     "use-debounce": "^10.0.2",
     "vaul": "^0.9.1",
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
+    "@tanstack/eslint-plugin-query": "^5.51.15",
+    "@types/node": "^20.14.15",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "eslint": "^8.57.0",
     "eslint-config-next": "14.2.5",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "postcss": "^8.4.41",
+    "tailwindcss": "^3.4.9",
+    "typescript": "^5.5.4"
   }
 }
