@@ -9,9 +9,16 @@ import {
   DropdownMenuLabel,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { MoreHorizontal } from "lucide-react";
+import { CalendarDays, MoreHorizontal } from "lucide-react";
 import { formatDistanceToNow } from "date-fns";
 import { Post } from "@/data/posts";
+import {
+  HoverCard,
+  HoverCardContent,
+  HoverCardTrigger,
+} from "@/components/ui/hover-card";
+import CommunityHoverCard from "./CommunityHoverCard";
+import { af, when } from "@/lib/utils";
 
 const PostHeader: React.FC<{ post: Post; isDetailsPage?: boolean }> = ({
   post,
@@ -20,30 +27,32 @@ const PostHeader: React.FC<{ post: Post; isDetailsPage?: boolean }> = ({
   return (
     <div key={post.id}>
       <div className="flex items-center justify-start">
-        <Avatar className={!isDetailsPage ? 'h-7 w-7': 'h-10 w-10'}>
-          <AvatarImage src={post.community.src} alt={post.community.alt} />
-          <AvatarFallback>{post.community.fallback}</AvatarFallback>
-        </Avatar>
+        <CommunityHoverCard community={post.community}>
+          <Avatar className={!isDetailsPage ? "h-7 w-7" : "h-10 w-10"}>
+          <AvatarImage src={post.community?.icon_img} alt={post?.community?.name} />
+          <AvatarFallback>{af(post.community?.name)}</AvatarFallback>
+          </Avatar>
+        </CommunityHoverCard>
 
         <div className="flex-1 space-y-2">
           <div className="flex items-center justify-between">
-            <div className="flex items-center text-xs ">
-              <div className="ml-2 ">
-                <div className="flex">
-                  <div className="font-bold">{post.community.text}</div>
-                  <div className="text-muted-foreground mx-3">
-                    {formatDistanceToNow(new Date(post.createdAt ?? ""), {
-                      addSuffix: false,
-                    })}
+            <CommunityHoverCard community={post?.community}>
+              <div className="flex items-center text-xs ">
+                <div className="ml-2 ">
+                  <div className="flex">
+                    <div className="font-bold">{post?.community?.name}</div>
+                    <div className="text-muted-foreground mx-3">
+                      { when(post?.created_at) }
+                    </div>
                   </div>
-              
+                  {isDetailsPage && (
+                    <div className="text-muted-foreground">
+                      {post?.community?.title}
+                    </div>
+                  )}
                 </div>
-                {isDetailsPage &&  <div className="text-muted-foreground">
-                  {post.community.title}
-                </div> }
-               
               </div>
-            </div>
+            </CommunityHoverCard>
 
             <div className="flex items-center">
               {!isDetailsPage && (
@@ -62,7 +71,6 @@ const PostHeader: React.FC<{ post: Post; isDetailsPage?: boolean }> = ({
                   </Button>
                 </DropdownMenuTrigger>
                 <DropdownMenuContent align="end">
-                 
                   <DropdownMenuItem>Save</DropdownMenuItem>
                   <DropdownMenuItem>Hide</DropdownMenuItem>
                   <DropdownMenuItem>Report</DropdownMenuItem>
