@@ -1,12 +1,21 @@
+import { users } from "@/data/users";
+import { communities } from "./communities";
+import { posts } from "@/data/posts";
+
 export const comments = [
   {
-    author: { name: "LaceLoverSings", avatar: "/path/to/avatar1.png" },
+    post: posts[0],
+    community: communities[0],
+    author: users[Math.floor(Math.random() * 6)],
     time: "7h ago",
     content:
       'I thought it would be a simple "will you marry me" bro did the whole speech in Chinese omg!',
     upvotes: 6800,
+    created_at: "2023-03-01T12:00:00Z",
     replies: [
       {
+        post: posts[0],
+        community: communities[0],
         author: { name: "Meatwood_Flak", avatar: "/path/to/avatar2.png" },
         time: "6h ago",
         content:
@@ -14,12 +23,16 @@ export const comments = [
         upvotes: 2300,
         replies: [
           {
-            author: { name: "Celestial_Crook", avatar: "/path/to/avatar4.png" },
+            post: posts[0],
+            community: communities[0],
+            author: users[Math.floor(Math.random() * 6)],
             time: "4h ago",
             content: "To be fair, Cantonese is A LOT harder than Mandarin.",
             upvotes: 58,
             replies: [
               {
+                post: posts[0],
+                community: communities[0],
                 author: {
                   name: "Celestial_Crook",
                   avatar: "/path/to/avatar4.png",
@@ -30,16 +43,21 @@ export const comments = [
                 replies: [
                   // Further nested replies if any
                 ],
+                created_at: "2023-03-01T12:00:00Z",
               },
             ],
           },
           {
-            author: { name: "Celestial_Crook", avatar: "/path/to/avatar4.png" },
+            post: posts[0],
+            community: communities[0],
+            author: users[Math.floor(Math.random() * 6)],
             time: "4h ago",
             content: "To be fair, Cantonese is A LOT harder than Mandarin.",
             upvotes: 58,
             replies: [
               {
+                post: posts[0],
+                community: communities[0],
                 author: {
                   name: "Celestial_Crook",
                   avatar: "/path/to/avatar4.png",
@@ -50,48 +68,60 @@ export const comments = [
                 replies: [
                   // Further nested replies if any
                 ],
+                created_at: "2023-03-01T12:00:00Z",
               },
             ],
+            created_at: "2023-03-01T12:00:00Z",
           },
         ],
       },
       {
-        author: { name: "Celestial_Crook", avatar: "/path/to/avatar4.png" },
+        post: posts[0],
+        community: communities[0],
+        author: users[Math.floor(Math.random() * 6)],
         time: "4h ago",
         content: "To be fair, Cantonese is A LOT harder than Mandarin.",
         upvotes: 58,
         replies: [
           // Further nested replies if any
         ],
+        created_at: "2023-03-01T12:00:00Z",
       },
       {
-        author: { name: "Celestial_Crook", avatar: "/path/to/avatar4.png" },
+        post: posts[0],
+        community: communities[0],
+        author: users[Math.floor(Math.random() * 6)],
         time: "4h ago",
         content: "To be fair, Cantonese is A LOT harder than Mandarin.",
         upvotes: 58,
         replies: [
           {
-            author: { name: "Celestial_Crook", avatar: "/path/to/avatar4.png" },
+            author: users[Math.floor(Math.random() * 6)],
             time: "4h ago",
             content: "To be fair, Cantonese is A LOT harder than Mandarin.",
             upvotes: 58,
             replies: [
               // Further nested replies if any
             ],
+            created_at: "2023-03-01T12:00:00Z",
           },
         ],
       },
     ],
   },
   {
-    author: { name: "UberShrew", avatar: "/path/to/avatar3.png" },
+    post: posts[0],
+    community: communities[0],
+    author: users[Math.floor(Math.random() * 6)],
     time: "6h ago",
     content:
       "Ha my wife would've shut my ass down after the 2nd sentence before she keeled over from my pronunciation if I had tried this in Cantonese. I probably had to say lou po 50 times before she found it acceptable. I still plan on learning it at some point if I can ever find a legit lesson plan. I have to talk like a robot because if I accidentally let any emotion into it the word comes out wrong and that's even after a year of taking mandarin.",
     upvotes: 77,
     replies: [
       {
-        author: { name: "Celestial_Crook", avatar: "/path/to/avatar4.png" },
+        post: posts[0],
+        community: communities[0],
+        author: users[Math.floor(Math.random() * 6)],
         time: "4h ago",
         content:
           "To be fair, Cantonese is A LOT harder than Mandarin. To be fair, Cantonese is A LOT harder than Mandarin. To be fair, Cantonese is A LOT harder than Mandarin.",
@@ -99,25 +129,32 @@ export const comments = [
         replies: [
           // Further nested replies if any
         ],
+        created_at: "2023-03-01T12:00:00Z",
       },
       {
-        author: { name: "Second Child", avatar: "/path/to/avatar4.png" },
+        post: posts[0],
+        community: communities[0],
+        author: users[Math.floor(Math.random() * 6)],
         time: "4h ago",
         content: "To be fair, Cantonese is A LOT harder than Mandarin.",
         upvotes: 58,
         replies: [
           // Further nested replies if any
         ],
+        created_at: "2023-03-01T12:00:00Z",
       },
       // More replies if any
     ],
   },
   {
-    author: { name: "AJ_Deadshow", avatar: "/path/to/avatar5.png" },
+    post: posts[0],
+    community: communities[0],
+    author: users[Math.floor(Math.random() * 6)],
     time: "6h ago",
     content:
       "I'm a man so normally I'm not allowed to get sappy, but this post is a whole sugar maple tree. The sappiest thing I've seen today.",
     upvotes: null,
     replies: [],
+    created_at: "2023-03-01T12:00:00Z",
   },
 ];
