@@ -1,4 +1,5 @@
 import { posts as postsData } from "@/data/posts";
 import { comments as commentsData } from "@/data/comments";
-
-export { postsData, commentsData };
+import { communities as communitiesData } from "@/data/communities";
+import { users as userData } from "@/data/users";
+export { postsData, commentsData, userData, communitiesData };
